# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SchoolBooks API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://schoo-api.onrender.com/api",
            "description": "API Server"
        },
        {
            "url": "http://localhost:8081/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get user's wishlist items",
                "description": "Returns current user's wishlist items with full product details",
                "operationId": "716344800c279861c61bfee7c75d624d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Wishlist"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/add-product": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add product to wishlist",
                "description": "Add a product to the current user's wishlist",
                "operationId": "cb8756f796670709bca3f7b5ef0f8746",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product added to wishlist successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wishlist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Product already in wishlist or validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/remove-product/{product_id}": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove product from wishlist",
                "description": "Remove a product from the current user's wishlist",
                "operationId": "28b3653ebb088b376280f7c179b4dda8",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID of the product to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from wishlist successfully"
                    },
                    "404": {
                        "description": "Product not found in wishlist"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/clear": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Clear user's wishlist",
                "description": "Remove all items from the current user's wishlist",
                "operationId": "2a97b8f037b7ae56bfc12f0acbf1a9b2",
                "responses": {
                    "200": {
                        "description": "Wishlist cleared successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/count": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get wishlist items count",
                "description": "Get the total number of items in the current user's wishlist",
                "operationId": "3ae10d2cd6653376a20207524753df02",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": "5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/{id}": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get specified wishlist item",
                "description": "Returns specified wishlist item (only if it belongs to current user)",
                "operationId": "74ea2848ee0ce4f468672f8a8faa4121",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the wishlist item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wishlist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wishlist item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the product",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Smartphone"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "A high-end smartphone with advanced features."
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "format": "float",
                        "example": "599.99"
                    },
                    "seller_id": {
                        "description": "ID of the seller who listed the product",
                        "type": "integer",
                        "example": "5"
                    },
                    "buyer_id": {
                        "description": "ID of the user who bought the product",
                        "type": "integer",
                        "example": "3",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status of the product (selling, sold, bought)",
                        "type": "string",
                        "enum": [
                            "selling",
                            "sold",
                            "bought"
                        ],
                        "example": "selling"
                    },
                    "category_id": {
                        "description": "ID of the category this product belongs to",
                        "type": "integer",
                        "example": "3"
                    },
                    "image_url": {
                        "description": "URL to the product image",
                        "type": "string",
                        "example": "https://example.com/images/smartphone.jpg"
                    },
                    "created_at": {
                        "description": "Date and time when the product was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 03:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 03:30:00"
                    },
                    "year_of_publication": {
                        "description": "Year the book was published",
                        "type": "integer",
                        "example": "2020"
                    },
                    "state_id": {
                        "description": "ID of the book's condition state",
                        "type": "integer",
                        "example": "1"
                    },
                    "author": {
                        "description": "Author of the book",
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "full_name": {
                        "description": "Full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "description": "Password (hashed)",
                        "type": "string",
                        "format": "password",
                        "example": "$2y$10$..."
                    },
                    "phone_number": {
                        "description": "Phone number",
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "description": "Email verification timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Wishlist": {
                "title": "Wishlist",
                "description": "Wishlist model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the wishlist item",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID associated with the wishlist item",
                        "type": "integer",
                        "example": "1"
                    },
                    "product_id": {
                        "description": "Product ID associated with the wishlist item",
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "description": "Date and time when the wishlist item was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 10:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the wishlist item was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 10:30:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}